+++ rastervision_pytorch_learner/rastervision/pytorch_learner/semantic_segmentation_learner.py
@@ -1,46 +1,95 @@
 import warnings
 warnings.filterwarnings('ignore')  # noqa
-from os.path import join, isdir, basename
+
+from typing import Union, IO, Callable, List, Iterable, Optional, Tuple
+from os.path import join, isdir
+from pathlib import Path
+
 import logging
-import glob
 
 import numpy as np
 import matplotlib
+from matplotlib import pyplot as plt
+import matplotlib.patches as mpatches
 matplotlib.use('Agg')  # noqa
-import torch
-from torch.utils.data import Dataset, ConcatDataset
-import torch.nn.functional as F
-from torchvision import models
 from PIL import Image
 
+import torch
+from torch import nn
+import torch.nn.functional as F
+from torch.utils.data import Dataset, ConcatDataset
+from torchvision import models
+
+from rastervision.pipeline.config import ConfigError
 from rastervision.pytorch_learner.learner import Learner
 from rastervision.pytorch_learner.utils import (
-    compute_conf_mat_metrics, compute_conf_mat, color_to_triple)
+    compute_conf_mat_metrics, compute_conf_mat, color_to_triple, SplitTensor,
+    Parallel, AddTensors)
+from rastervision.pipeline.file_system import make_dir
 
 log = logging.getLogger(__name__)
 
 
+def load_np_normalized(path: Union[IO, str, Path]) -> np.ndarray:
+    arr = np.load(path)
+    dtype = arr.dtype
+    if np.issubdtype(dtype, np.unsignedinteger):
+        max_val = np.iinfo(dtype).max
+        arr = arr.astype(np.float32) / max_val
+    return arr
+
+
 class SemanticSegmentationDataset(Dataset):
-    def __init__(self, data_dir, transform=None):
-        self.data_dir = data_dir
-        self.img_paths = glob.glob(join(data_dir, 'img', '*.png'))
+    def __init__(self,
+                 data_dir: str,
+                 img_fmt: str = 'png',
+                 label_fmt: str = 'png',
+                 transform: Callable = None):
+
+        self.data_dir = Path(data_dir)
+        img_dir = self.data_dir / 'img'
+        label_dir = self.data_dir / 'labels'
+
+        # collect image and label paths
+        self.img_paths = list(img_dir.glob(f'*.{img_fmt}'))
+        self.label_paths = [
+            label_dir / f'{p.stem}.{label_fmt}' for p in self.img_paths
+        ]
+
+        # choose image loading method based on format
+        if img_fmt.lower() in ('npy', 'npz'):
+            self.img_load_fn = load_np_normalized
+        else:
+            self.img_load_fn = lambda path: np.array(Image.open(path)) / 255.
+
+        # choose label loading method based on format
+        if label_fmt.lower() in ('npy', 'npz'):
+            self.label_load_fn = np.load
+        else:
+            self.label_load_fn = lambda path: np.array(Image.open(path))
+
         self.transform = transform
 
-    def __getitem__(self, ind):
-        img_path = self.img_paths[ind]
-        label_path = join(self.data_dir, 'labels', basename(img_path))
-        x = Image.open(img_path)
-        y = Image.open(label_path)
+    def __getitem__(self,
+                    ind: int) -> Tuple[torch.FloatTensor, torch.LongTensor]:
 
-        x = np.array(x)
-        y = np.array(y)
+        img_path = self.img_paths[ind]
+        label_path = self.label_paths[ind]
+
+        x = self.img_load_fn(img_path)
+        y = self.label_load_fn(label_path)
+
+        if x.ndim == 2:
+            # (h, w) --> (h, w, 1)
+            x = x[..., np.newaxis]
+
         if self.transform is not None:
             out = self.transform(image=x, mask=y)
             x = out['image']
             y = out['mask']
 
-        x = torch.tensor(x).permute(2, 0, 1).float() / 255.0
-        y = torch.tensor(y).long()
+        x = torch.from_numpy(x).permute(2, 0, 1).float()
+        y = torch.from_numpy(y).long()
 
         return (x, y)
 
@@ -49,7 +98,7 @@ class SemanticSegmentationDataset(Dataset):
 
 
 class SemanticSegmentationLearner(Learner):
-    def build_model(self):
+    def build_model(self) -> nn.Module:
         # TODO support FCN option
         pretrained = self.cfg.model.pretrained
         model = models.segmentation.segmentation._segm_resnet(
@@ -58,13 +107,57 @@ class SemanticSegmentationLearner(Learner):
             len(self.cfg.data.class_names),
             False,
             pretrained_backbone=pretrained)
+
+        input_channels = self.cfg.data.img_channels
+        old_conv = model.backbone.conv1
+
+        if input_channels == old_conv.in_channels:
+            return model
+
+        # these parameters will be the same for the new conv layer
+        old_conv_args = {
+            'out_channels': old_conv.out_channels,
+            'kernel_size': old_conv.kernel_size,
+            'stride': old_conv.stride,
+            'padding': old_conv.padding,
+            'dilation': old_conv.dilation,
+            'groups': old_conv.groups,
+            'bias': old_conv.bias
+        }
+
+        if not pretrained:
+            # simply replace the first conv layer with one with the
+            # correct number of input channels
+            new_conv = nn.Conv2d(in_channels=input_channels, **old_conv_args)
+            model.backbone.conv1 = new_conv
+            return model
+
+        if input_channels > old_conv.in_channels:
+            # insert a new conv layer parallel to the existing one
+            # and sum their outputs
+            new_conv_channels = input_channels - old_conv.in_channels
+            new_conv = nn.Conv2d(
+                in_channels=new_conv_channels, **old_conv_args)
+            model.backbone.conv1 = nn.Sequential(
+                # split input along channel dim
+                SplitTensor((old_conv.in_channels, new_conv_channels), dim=1),
+                # each split goes to its respective conv layer
+                Parallel(old_conv, new_conv),
+                # sum the parallel outputs
+                AddTensors())
+        else:
+            raise ConfigError(
+                (f'Fewer input channels ({input_channels}) than what'
+                 f'the pretrained model expects ({old_conv.in_channels})'))
+
         return model
 
-    def _get_datasets(self, uri):
+    def _get_datasets(self, uri: Union[str, List[str]]):
         cfg = self.cfg
 
         data_dirs = self.unzip_data(uri)
         transform, aug_transform = self.get_data_transforms()
+        img_fmt, label_fmt = cfg.data.img_format, cfg.data.label_format
 
         train_ds, valid_ds, test_ds = [], [], []
         for data_dir in data_dirs:
@@ -72,22 +165,27 @@ class SemanticSegmentationLearner(Learner):
             valid_dir = join(data_dir, 'valid')
 
             if isdir(train_dir):
-                if cfg.overfit_mode:
-                    train_ds.append(
-                        SemanticSegmentationDataset(
-                            train_dir, transform=transform))
-                else:
-                    train_ds.append(
-                        SemanticSegmentationDataset(
-                            train_dir, transform=aug_transform))
+                tf = transform if cfg.overfit_mode else aug_transform
+                ds = SemanticSegmentationDataset(
+                    train_dir,
+                    img_fmt=img_fmt,
+                    label_fmt=label_fmt,
+                    transform=tf)
+                train_ds.append(ds)
 
             if isdir(valid_dir):
                 valid_ds.append(
                     SemanticSegmentationDataset(
-                        valid_dir, transform=transform))
+                        valid_dir,
+                        img_fmt=img_fmt,
+                        label_fmt=label_fmt,
+                        transform=transform))
                 test_ds.append(
                     SemanticSegmentationDataset(
-                        valid_dir, transform=transform))
+                        valid_dir,
+                        img_fmt=img_fmt,
+                        label_fmt=label_fmt,
+                        transform=transform))
 
         train_ds, valid_ds, test_ds = \
             ConcatDataset(train_ds), ConcatDataset(valid_ds), ConcatDataset(test_ds)
@@ -129,22 +227,109 @@ class SemanticSegmentationLearner(Learner):
     def prob_to_pred(self, x):
         return x.argmax(1)
 
-    def plot_xyz(self, ax, x, y, z=None):
-        x = x.permute(1, 2, 0)
-        if x.shape[2] == 1:
-            x = torch.cat([x for _ in range(3)], dim=2)
-        ax.imshow(x)
-        ax.axis('off')
+    def plot_batch(self,
+                   x: torch.Tensor,
+                   y: Union[torch.Tensor, np.ndarray],
+                   output_path: str,
+                   z: Optional[torch.Tensor] = None) -> None:
+        """Plot a whole batch in a grid using plot_xyz.
 
-        labels = z if z is not None else y
-        colors = [color_to_triple(c) for c in self.cfg.data.class_colors]
-        colors = [tuple([_c / 255 for _c in c]) for c in colors]
+        Args:
+            x: batch of images
+            y: ground truth labels
+            output_path: local path where to save plot image
+            z: optional predicted labels
+        """
+        batch_sz, c, h, w = x.shape
+        channel_groups = self.cfg.data.channel_display_groups
+
+        nrows = batch_sz
+        # one col for each group + 1 for labels + 1 for predictions
+        ncols = len(channel_groups) + 1
+        if z is not None:
+            ncols += 1
+
+        fig, axes = plt.subplots(
+            nrows=nrows,
+            ncols=ncols,
+            constrained_layout=True,
+            figsize=(3 * ncols, 3 * nrows))
+
+        for i in range(batch_sz):
+            ax = (fig, axes[i])
+            if z is None:
+                self.plot_xyz(ax, x[i], y[i])
+            else:
+                self.plot_xyz(ax, x[i], y[i], z=z[i])
+
+        make_dir(output_path, use_dirname=True)
+        plt.savefig(output_path, bbox_inches='tight')
+        plt.close()
+
+    def plot_xyz(self,
+                 ax: Iterable,
+                 x: torch.Tensor,
+                 y: Union[torch.Tensor, np.ndarray],
+                 z: Optional[torch.Tensor] = None) -> None:
+
+        channel_groups = self.cfg.data.channel_display_groups
+
+        # make subplot titles
+        if not isinstance(channel_groups, dict):
+            channel_groups = {
+                f'Channels: {[*chs]}': chs
+                for chs in channel_groups
+            }
+
+        fig, ax = ax
+        img_axes = ax[:len(channel_groups)]
+        label_ax = ax[len(channel_groups)]
+
+        # (c, h, w) --> (h, w, c)
+        x = x.permute(1, 2, 0)
+
+        # plot input image(s)
+        for (title, chs), ch_ax in zip(channel_groups.items(), img_axes):
+            im = x[..., chs]
+            if len(chs) == 1:
+                im = im.expand(-1, -1, 3)
+            elif len(chs) == 2:
+                h, w, _ = x.shape
+                third_channel = torch.full((h, w, 1), fill_value=.5)
+                im = torch.cat((im, third_channel), dim=-1)
+            ch_ax.imshow(im)
+            ch_ax.set_title(title)
+            ch_ax.set_xticks([])
+            ch_ax.set_yticks([])
+
+        class_colors = self.cfg.data.class_colors
+        colors = [color_to_triple(c) for c in class_colors]
+        colors = np.array(colors) / 255.
         cmap = matplotlib.colors.ListedColormap(colors)
-        labels = labels.numpy()
-        ax.imshow(
-            labels,
-            alpha=0.4,
-            vmin=0,
-            vmax=len(colors),
-            cmap=cmap,
-            interpolation='none')
+
+        # plot labels
+        label_ax.imshow(
+            y, vmin=0, vmax=len(colors), cmap=cmap, interpolation='none')
+        label_ax.set_title(f'Ground truth labels')
+        label_ax.set_xticks([])
+        label_ax.set_yticks([])
+
+        # plot predictions
+        if z is not None:
+            pred_ax = ax[-1]
+            pred_ax.imshow(
+                z, vmin=0, vmax=len(colors), cmap=cmap, interpolation='none')
+            pred_ax.set_title(f'Predicted labels')
+            pred_ax.set_xticks([])
+            pred_ax.set_yticks([])
+
+        # add a legend to the rightmost subplot
+        class_names = self.cfg.data.class_names
+        legend_items = [
+            mpatches.Patch(facecolor=col, edgecolor='black', label=name)
+            for col, name in zip(colors, class_names)
+        ]
+        ax[-1].legend(
+            handles=legend_items,
+            loc='center right',
+            bbox_to_anchor=(1.8, 0.5))