+++ rastervision_pytorch_learner/rastervision/pytorch_learner/semantic_segmentation_learner_config.py
@@ -1,4 +1,6 @@
 from enum import Enum
+from typing import Union, Optional
+from pydantic import PositiveInt
 
 from rastervision.pipeline.config import register_config, Field, validator
 from rastervision.pytorch_learner.learner_config import (
@@ -13,6 +15,37 @@ class SemanticSegmentationDataFormat(Enum):
 @register_config('semantic_segmentation_data')
 class SemanticSegmentationDataConfig(DataConfig):
     data_format: SemanticSegmentationDataFormat = SemanticSegmentationDataFormat.default
+
+    img_channels: PositiveInt = Field(
+        3, description='The number of channels of the training images.')
+
+    img_format: Optional[str] = Field(
+        None, description='The filetype of the training images.')
+    label_format: str = Field(
+        'png', description='The filetype of the training labels.')
+
+    channel_display_groups: Optional[Union[dict, list, tuple]] = Field(
+        None,
+        description=
+        ('Groups of image channels to display together as a subplot '
+         'when plotting the data and predictions. '
+         'Can be a list or tuple of groups (e.g. [(0, 1, 2), (3,)]) or a dict '
+         'containing title-to-group mappings '
+         '(e.g. {"RGB": [0, 1, 2], "IR": [3]}), '
+         'where each group is a list or tuple of channel indices and title '
+         'is a string that will be used as the title of the subplot '
+         'for that group.'))
+
+    def update(self, **kwargs):
+        super().update()
+
+        if self.img_format is None:
+            self.img_format = 'png' if self.img_channels == 3 else 'npy'
+
+        if self.channel_display_groups is None:
+            self.channel_display_groups = {
+                'Input': tuple(range(self.img_channels))
+            }
 
 
 @register_config('semantic_segmentation_model')